% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMOEXData.R
\name{getMOEXData}
\alias{getMOEXData}
\title{download market data from MOEX Exchange using ISS engine}
\usage{
getMOEXData(ticker, from, to, interval = 1)
}
\arguments{
\item{ticker}{ticker name}

\item{from}{start date}

\item{to}{end date}

\item{interval}{24 (1day), 60 (1hour), 10 (10min), 1 (1min)}
}
\value{
data.table
}
\description{
download market data from MOEX Exchange using ISS engine
}
\examples{
#ETF
getMOEXData(ticker = 'FXCN',from='2020-09-01',to='2020-09-10', interval = 1)
#Futures
getMOEXData(ticker = 'SiZ0',from='2020-09-01',to='2020-09-10', interval = 24)
#Stock
getMOEXData(ticker = 'SBER',from='2006-01-01',to='2020-09-10', interval = 24)
#Option
getMOEXData(ticker = 'Si70000BL0',from='2020-09-01',to='2020-09-10', interval = 24)
#Currency
getMOEXData(ticker = 'USD000000TOD',from='2020-09-01',to='2020-09-10', interval = 24)
#Depository receipts
getMOEXData(ticker = 'FIVE',from='2020-09-01',to='2020-09-10', interval = 24)
#Bonds
getMOEXData(ticker = 'RU000A101FG8',from='2020-09-01',to='2020-09-10', interval = 24)

}
