% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMOEXData.R
\name{getMOEXData}
\alias{getMOEXData}
\title{download market data from MOEX Exchange using ISS engine}
\usage{
getMOEXData(
  ticker = "SBER",
  from = Sys.Date() - 2,
  to = Sys.Date(),
  period = "1min"
)
}
\arguments{
\item{ticker}{ticker name}

\item{from}{start date}

\item{to}{end date}

\item{period}{candle period \code{day, hour, 10min, 1min}}
}
\value{
data.table
}
\description{
download market data from MOEX Exchange using ISS engine
}
\examples{
#ETF
getMOEXData(ticker = 'FXCN',from='2020-09-01',to='2020-09-10', period = 'day')
#Futures
getMOEXData(ticker = 'SiZ0',from='2020-09-01',to='2020-09-10', period = 'day')
#Stock
getMOEXData(ticker = 'SBER',from='2006-01-01',to='2020-09-10', period = 'day')
#Option
getMOEXData(ticker = 'Si70000BL0',from='2020-09-01',to='2020-09-10', period = 'day')
#Currency
getMOEXData(ticker = 'USD000000TOD',from='2020-09-01',to='2020-09-10', period = 'day')
#Depository receipts
getMOEXData(ticker = 'FIVE',from='2020-09-01',to='2020-09-10', period = 'day')
#Bonds
getMOEXData(ticker = 'RU000A101FG8',from='2020-09-01',to='2020-09-10', period = 'day')
}
